fix: Fix `nextLong()`
This fixes compatibility with BetterX, and eliminates any differences between vanilla and Faster Random, meaning Faster Random now has 1:1 parity with vanilla (maybe except for the MC-239059 fix).

+++ src/main/java/com/github/anopensaucedev/fasterrandom/util/math/random/RandomGeneratorRandom.java
@@ -59,7 +59,7 @@ public class RandomGeneratorRandom implements BaseRandom {
 
 	@Override
 	public long nextLong() {
-		return randomGenerator.nextInt();
+		return randomGenerator.nextLong();
 	}
 
 	@Override
